requirements.txt
paho-mqtt
pixel-ring


